# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Enterprise Real Estate ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /unit-price-lists:
    get:
      operationId: listUnitPriceLists
      tags: [unitPriceLists]
      summary: List unit-price-lists
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: pageSize, in: query, required: false, schema: { type: integer } }
        - { name: search, in: query, required: false, schema: { type: string } }
        - { name: companyId, in: query, required: false, schema: { type: string } }
        - { name: projectId, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: UnitPriceLists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitPriceListListResponse"
    post:
      operationId: createUnitPriceList
      tags: [unitPriceLists]
      summary: Create a UnitPriceList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitPriceListInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitPriceList"
  /unit-price-lists/{id}:
    get:
      operationId: getUnitPriceList
      tags: [unitPriceLists]
      summary: Get a UnitPriceList
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: UnitPriceList
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitPriceList"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateUnitPriceList
      tags: [unitPriceLists]
      summary: Update a UnitPriceList
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitPriceListUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitPriceList"
    delete:
      operationId: deleteUnitPriceList
      tags: [unitPriceLists]
      summary: Soft-delete a UnitPriceList
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /unit-pricing:
    get:
      operationId: listUnitPricings
      tags: [unitPricing]
      summary: List unit-pricing
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: pageSize, in: query, required: false, schema: { type: integer } }
        - { name: search, in: query, required: false, schema: { type: string } }
        - { name: companyId, in: query, required: false, schema: { type: string } }
        - { name: priceListId, in: query, required: false, schema: { type: string } }
        - { name: unitId, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: UnitPricings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitPricingListResponse"
    post:
      operationId: createUnitPricing
      tags: [unitPricing]
      summary: Create a UnitPricing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitPricingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitPricing"
  /unit-pricing/{id}:
    get:
      operationId: getUnitPricing
      tags: [unitPricing]
      summary: Get a UnitPricing
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: UnitPricing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitPricing"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateUnitPricing
      tags: [unitPricing]
      summary: Update a UnitPricing
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitPricingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitPricing"
    delete:
      operationId: deleteUnitPricing
      tags: [unitPricing]
      summary: Soft-delete a UnitPricing
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /unit-discounts:
    get:
      operationId: listUnitDiscounts
      tags: [unitDiscounts]
      summary: List unit-discounts
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: pageSize, in: query, required: false, schema: { type: integer } }
        - { name: search, in: query, required: false, schema: { type: string } }
        - { name: companyId, in: query, required: false, schema: { type: string } }
        - { name: discountType, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: UnitDiscounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitDiscountListResponse"
    post:
      operationId: createUnitDiscount
      tags: [unitDiscounts]
      summary: Create a UnitDiscount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitDiscountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitDiscount"
  /unit-discounts/{id}:
    get:
      operationId: getUnitDiscount
      tags: [unitDiscounts]
      summary: Get a UnitDiscount
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: UnitDiscount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitDiscount"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateUnitDiscount
      tags: [unitDiscounts]
      summary: Update a UnitDiscount
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitDiscountUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitDiscount"
    delete:
      operationId: deleteUnitDiscount
      tags: [unitDiscounts]
      summary: Soft-delete a UnitDiscount
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /unit-availability:
    get:
      operationId: getUnitAvailability
      tags: [unitManagement]
      summary: Unit availability summary
      parameters:
        - { name: companyId, in: query, required: false, schema: { type: string } }
        - { name: projectId, in: query, required: false, schema: { type: string } }
        - { name: buildingId, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: Availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitAvailability"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    OkResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
...
    UnitPriceList:
      type: object
      required: [id, companyId, code, name, nameAr, isActive, createdAt]
      properties:
        id: { type: string }
        companyId: { type: string }
        code: { type: string }
        name: { type: string }
        nameAr: { type: string }
        projectId: { type: ["string", "null"] }
        effectiveFrom: { type: ["string", "null"] }
        effectiveTo: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string }
    UnitPriceListInput:
      type: object
      required: [companyId, code, name, nameAr]
      properties:
        companyId: { type: string }
        code: { type: string }
        name: { type: string }
        nameAr: { type: string }
        projectId: { type: string }
        effectiveFrom: { type: string }
        effectiveTo: { type: string }
    UnitPriceListUpdate:
      type: object
      properties:
        companyId: { type: string }
        code: { type: string }
        name: { type: string }
        nameAr: { type: string }
        projectId: { type: string }
        effectiveFrom: { type: string }
        effectiveTo: { type: string }
    UnitPriceListListResponse:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/UnitPriceList" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    UnitPricing:
      type: object
      required: [id, companyId, priceListId, unitId, price, isActive, createdAt]
      properties:
        id: { type: string }
        companyId: { type: string }
        priceListId: { type: string }
        unitId: { type: string }
        price: { type: string }
        effectiveDate: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string }
    UnitPricingInput:
      type: object
      required: [companyId, priceListId, unitId]
      properties:
        companyId: { type: string }
        priceListId: { type: string }
        unitId: { type: string }
        price: { type: string }
        effectiveDate: { type: string }
    UnitPricingUpdate:
      type: object
      properties:
        companyId: { type: string }
        priceListId: { type: string }
        unitId: { type: string }
        price: { type: string }
        effectiveDate: { type: string }
    UnitPricingListResponse:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/UnitPricing" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    UnitDiscount:
      type: object
      required: [id, companyId, code, name, nameAr, discountType, discountValue, isActive, createdAt]
      properties:
        id: { type: string }
        companyId: { type: string }
        code: { type: string }
        name: { type: string }
        nameAr: { type: string }
        discountType: { type: string }
        discountValue: { type: string }
        validFrom: { type: ["string", "null"] }
        validTo: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string }
    UnitDiscountInput:
      type: object
      required: [companyId, code, name, nameAr]
      properties:
        companyId: { type: string }
        code: { type: string }
        name: { type: string }
        nameAr: { type: string }
        discountType: { type: string }
        discountValue: { type: string }
        validFrom: { type: string }
        validTo: { type: string }
    UnitDiscountUpdate:
      type: object
      properties:
        companyId: { type: string }
        code: { type: string }
        name: { type: string }
        nameAr: { type: string }
        discountType: { type: string }
        discountValue: { type: string }
        validFrom: { type: string }
        validTo: { type: string }
    UnitDiscountListResponse:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/UnitDiscount" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    UnitAvailability:
      type: object
      required: [total, available, reserved, sold, other]
      properties:
        total: { type: integer }
        available: { type: integer }
        reserved: { type: integer }
        sold: { type: integer }
        other: { type: integer }
...