# lib/api-zod/src/generated/types/unitPriceListUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Enterprise Real Estate ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface UnitPriceListUpdate {
  companyId?: string;
  code?: string;
  name?: string;
  nameAr?: string;
  projectId?: string;
  effectiveFrom?: string;
  effectiveTo?: string;
}